+++ src/server/api/mutations/index.js
@@ -1,15 +1,11 @@
-import {
-  sendMessage
-} from "./sendMessage";
-import {
-  bulkSendMessages
-} from "./bulkSendMessages";
-import {
-  findNewCampaignContact
-} from "./findNewCampaignContact";
+import { sendMessage } from "./sendMessage";
+import { bulkSendMessages } from "./bulkSendMessages";
+import { findNewCampaignContact } from "./findNewCampaignContact";
+import { buyPhoneNumbers } from "./buyPhoneNumbers";
 
 export {
   sendMessage,
   bulkSendMessages,
+  buyPhoneNumbers,
   findNewCampaignContact
 };
