+++ src/server/api/schema.js
@@ -10,8 +10,7 @@ import {
   dispatchContactIngestLoad,
   exportCampaign,
   importScript,
-  loadCampaignCache,
-  buyPhoneNumbers
+  loadCampaignCache
 } from "../../workers/jobs";
 import { getIngestMethod } from "../../integrations/contact-loaders";
 import {
@@ -65,9 +64,9 @@ import Twilio from "twilio";
 import {
   sendMessage,
   bulkSendMessages,
-  findNewCampaignContact
+  findNewCampaignContact,
+  buyPhoneNumbers
 } from "./mutations";
-import serviceMap from "./lib/services";
 
 const uuidv4 = require("uuid").v4;
 const JOBS_SAME_PROCESS = !!(
@@ -312,6 +311,7 @@ async function updateInteractionSteps(
 
 const rootMutations = {
   RootMutation: {
+    buyPhoneNumbers,
     userAgreeTerms: async (_, { userId }, { user, loaders }) => {
       if (user.id === Number(userId)) {
         return user.terms ? user : null;
@@ -1299,52 +1299,6 @@ const rootMutations = {
         .where("id", id)
         .update({ is_deleted: true });
       return { id };
-    },
-    buyPhoneNumbers: async (
-      _,
-      { organizationId, areaCode, limit, addToOrganizationMessagingService },
-      { loaders, user }
-    ) => {
-      await accessRequired(user, organizationId, "OWNER");
-      const org = await loaders.organization.load(organizationId);
-      if (!getConfig("EXPERIMENTAL_PHONE_INVENTORY", org, { truthy: true })) {
-        throw new Error("Phone inventory management is not enabled");
-      }
-      const serviceName = getConfig("DEFAULT_SERVICE", org);
-      const service = serviceMap[serviceName];
-      if (!service || !service.hasOwnProperty("buyNumbersInAreaCode")) {
-        throw new Error(
-          `Service ${serviceName} does not support phone number buying`
-        );
-      }
-
-      let messagingServiceSid;
-      if (addToOrganizationMessagingService) {
-        const msgSrv = JSON.parse(org.features || "{}")
-          .TWILIO_MESSAGE_SERVICE_SID;
-        if (serviceName !== "twilio" || !msgSrv) {
-          throw new Error(
-            "This organization is not configured to use its own Twilio Messaging Service"
-          );
-        }
-        messagingServiceSid = msgSrv;
-      }
-      const job = await JobRequest.save({
-        queue_name: `${organizationId}:buy_phone_numbers`,
-        organization_id: organizationId,
-        job_type: "buy_phone_numbers",
-        locks_queue: false,
-        assigned: JOBS_SAME_PROCESS,
-        payload: JSON.stringify({
-          areaCode,
-          limit,
-          messagingServiceSid
-        })
-      });
-      if (JOBS_SAME_PROCESS) {
-        buyPhoneNumbers(job);
-      }
-      return job;
     }
   }
 };
